% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversityTests.R
\name{allele.alphaContrastTest}
\alias{allele.alphaContrastTest}
\title{Test for difference in alpha diversity between two allele diversity datasets}
\usage{
allele.alphaContrastTest(lst.a, lst.b, zero.var.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"))
}
\arguments{
\item{lst.a}{Allele diveristy dataset, a list, one entry per locus, of
              site x allele counts. Each table must have the same format.}

\item{lst.b}{Allele diveristy dataset, a list, one entry per locus, of
              site x allele counts. Each table must have the same format.}

\item{\dots}{Additional parameters}
}
\value{
Fill this in
}
\description{
From Sork et al., extensions of those from Scofield et al. 2012 American
Naturalist 180(6) 719-732, http://www.jstor.org/stable/10.1086/668202).
}
\examples{
# For comparing allele diversity between two different samples:

  dat1 <- readGenalex("file-of-genotypes-sample-1.txt")
  dat2 <- readGenalex("file-of-genotypes-sample-2.txt")
  gt1 <- allele.createTableList(dat1)
  gt2 <- allele.createTableList(dat2)
  alpha.contrast <- allele.alphaContrastTest(gt1, gt2)
}
\author{
Douglas G. Scofield
}
\seealso{
\code{\link{}}
}

