% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversityTests.R
\name{allele.alphaDiversityTest}
\alias{allele.alphaDiversityTest}
\title{Test for alpha diversity difference between sites}
\usage{
allele.alphaDiversityTest(lst, zero.var.adjust = TRUE, n.resample = 10000,
  method = c("bootstrap", "permute"))
}
\arguments{
\item{lst}{Allele diversity dataset, a list, one entry per locus, of
            site x allele counts. Each table must have the same format.}

\item{\dots}{Additional parameters}
}
\value{
Fill this in
}
\description{
From Sork et al., extensions of those from Scofield et al. 2012 American
Naturalist 180(6) 719-732, http://www.jstor.org/stable/10.1086/668202).
}
\examples{
# For comparing allele diversity between two different samples:

  dat <- readGenalex("file-of-genotypes-sample-1.txt")
  gt <- allele.createTableList(dat)
  alpha.test <- allele.alphaDiversityTest(gt)
}
\author{
Douglas G. Scofield
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{alphaDiversityTest}}, \code{\link{diversity}}
}

