% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversity.R
\name{allele.diversity}
\alias{allele.diversity}
\title{allele.diversity}
\usage{
allele.diversity(lst)
}
\description{
allele.diversity
}
\examples{
# The workflow to calculate basic allelic diversity statistics:

  dat <- readGenalex("GenAlEx-format-file-of-genotypes.txt")
  gt <- allele.createTableList(dat)
  div <- allele.diversity(gt)
}

