% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversity.R
\name{allele.diversitySingleLocus}
\alias{allele.diversitySingleLocus}
\title{Calculate allelic diversity for a single locus}
\usage{
allele.diversitySingleLocus(tab)
}
\arguments{
\item{tab}{Table of site x allele counts, with row names being the site
names and column names being the names given to the individual alleles}
}
\value{
List of diversity estimates for the locus
}
\description{
Calculate allelic diversity for a single locus
}

