% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleGammaAccum.R
\name{allele.runGammaAccum}
\alias{allele.runGammaAccum}
\title{Perform gamma diversity accumulation on allele counts}
\usage{
allele.runGammaAccum(lst, accum.method = c("random", "proximity"),
  resample.method = c("permute", "bootstrap"), n.resample = 1000,
  gamma.method = c("r", "q.nielsen", "q"), distance.file = NULL, ...)
}
\arguments{
\item{lst}{List of site-by-source tables for loci, as produced by
\code{allele.createTableList}}

\item{accum.method}{\code{"random"} (default) or \code{"proximity"}.  If
\code{"proximity"} is used, then \code{distance.file} must be supplied
(see below).}

\item{resample.method}{\code{"permute"} (default) or \code{"bootstrap"},
whether to resample sites without (\code{"permute"}) or with
(\code{"bootstrap"}) replacement}

\item{n.resample}{Number of resamples for accumulation curve}

\item{gamma.method}{Calculate gamma using the \code{"r"} (default),
\code{"q"} or \code{"q.nielsen"} method, see Scofield et al. (2012).
The allele.* functions are all based on \code{"r"} versions and undefined
results may happen otherwise.}

\item{distance.file}{A file or data frame containing three columns of data,
with the header/column names being \code{pool}, \code{X}, and \code{Y},
containing the spatial locations of the seed pools named in the row names
of tab.  Only used if the \code{accum.method == "proximity"}.}
}
\description{
Several arguments control the method of accumulation and value of gamma
calculated.  Only the defaults have been tested; the others were developed
while exploring the data and must be considered experimental.  The result is
returned in a list, which may be passed to \code{\link{plotGammaAccum}} to
plot the result.
}
\note{
Only the defaults have been tested; other options were developed while
exploring the data and must be considered experimental.
}
\examples{
dat <- readGenalex("genotypes.txt")
lst <- allele.createTableList(dat)
allele.rga.result <- allele.runGammaAccum(lst)
plotGammaAccum(allele.rga.result)
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}

