% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diversityTests.R
\name{alphaContrastTest.3}
\alias{alphaContrastTest.3}
\title{Test for differences in alpha diversity between three data sets}
\usage{
alphaContrastTest.3(tab.a, tab.b, tab.c, zero.var.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"))
}
\arguments{
\item{tab.a}{First ite x soure table}

\item{tab.b}{Second site x soure table}

\item{tab.c}{Third site x soure table}

\item{zero.var.adjust}{Logical, whether to adjust zero-variance groups
with minimum value, see \code{\link{BLAHBLAH}}}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}
}
\value{
An \code{'htest_boot'} object with the result of the test. The
test also prints the result, so perhaps I should modify it to only
return the htest_boot object?  Or should it return a different object?
}
\description{
Test for differences in alpha diversity between three data sets
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{alphaContrastTest}}
}

