% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-allele_divtables.R, R/diversityTests-divtable.R
\name{alphaContrastTest.allele_divtables}
\alias{alphaContrastTest}
\alias{alphaContrastTest.allele_divtables}
\alias{alphaContrastTest.divtable}
\title{Test for differences in alpha diversity between two allele diversity datasets}
\usage{
\method{alphaContrastTest}{allele_divtables}(lst.a, lst.b,
  zero.var.adjust = TRUE, n.resample = 10000, method = c("bootstrap",
  "permute"), test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99,
  0.999))

\method{alphaContrastTest}{divtable}(tab.a, tab.b, zero.var.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999))
}
\arguments{
\item{zero.var.adjust}{Logical, whether to adjust zero-variance groups
with minimum value, see \code{\link{BLAHBLAH}}}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}

\item{tab.a}{First ite x soure table}

\item{tab.b}{Second site x soure table}
}
\value{
An \code{'htest_boot'} object with the result of the test. The
test also prints the result, so perhaps I should modify it to only
return the htest_boot object?  Or should it return a different object?
}
\description{
From Sork et al. (2015), extending methods developed in Scofield et al.
(2012).

Test for differences in alpha diversity between two data sets
}
\examples{
# For comparing allele diversity between two different samples:

# dat1 <- readGenalex("file-of-genotypes-sample-1.txt")
# dat2 <- readGenalex("file-of-genotypes-sample-2.txt")
# gt1 <- createAlleleTables(dat1)
# gt2 <- createAlleleTables(dat2)
# alpha.contrast <- alphaContrastTest(gt1, gt2)
}
\references{
Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (Submitted)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).

Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.

Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{alphaContrastTest}}

\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest3}}
}

