% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-allele_divtables.R, R/diversityTests-divtable.R
\name{alphaDiversityTest.allele_divtables}
\alias{alphaDiversityTest}
\alias{alphaDiversityTest.allele_divtables}
\alias{alphaDiversityTest.default}
\alias{alphaDiversityTest.divtable}
\title{Test for differences in alpha diversity among sites within a single data set}
\usage{
\method{alphaDiversityTest}{allele_divtables}(adt, zero.div.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999),
  ...)

\method{alphaDiversityTest}{divtable}(tab, zero.div.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999),
  ...)

\method{alphaDiversityTest}{default}(tab, ...)
}
\arguments{
\item{adt}{Allele diversity dataset of class
\code{\link{allele_divtables}}}

\item{zero.div.adjust}{Logical, if \code{TRUE} (the default), then groups
with 0 within-group diversity are assigned a minimum diversity which is
half the empirical diversity possible given the group size}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}

\item{tab}{Site-by-source table of class \code{\link{divtable}}}

\item{\dots}{Additional parameters}
}
\value{
An \code{diversity_test} object with the result of the test
}
\description{
The null hypothesis for this tests is that there is no difference in alpha
diversity between the sites represented in \code{tab} or \code{adt}.  The
initial (class \code{\link{divtable}}) version of this was described in
Scofield et al. (2012), while the allelic (class
\code{link{allele_divtables}} extension was described in Sort et al. (In
press).
}
\examples{
## Add example with class divtable
##
## Using allele diversity dataset of class allele_divtables.  Compare
## allele diversity between sites in the same sample:
##
library(readGenalex)
data(Qagr_pericarp_genotypes)
gt <- createAlleleTables(Qagr_pericarp_genotypes)
alpha.test <- alphaDiversityTest(gt)
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.

Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (In press)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).  \emph{Evolutionary Ecology}.
}
\seealso{
\code{\link{alphaContrastTest}}, \code{\link{alphaContrastTest3}}, \code{\link{print.diversity_test}}
}

