% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diversity-table.R
\name{diversity}
\alias{diversity}
\title{Calculate a collection of diversity statistics from a site x source matrix}
\usage{
diversity(tab)
}
\arguments{
\item{tab}{Table of counts in sites (rows) X sources (columns) format.
If the argument is not a matrix, it is converted to one.  Rows are
not reordered.}
}
\value{
A list is returned which contains three forms of diversity
statistics, each' returned as lists under separate named elements:
\itemize{
    \item \code{q}, based on squared frequencies, known to be biased at
          smaller sample sizes
    \item \code{r}, unbiased versions of \code{q} (Simpson, 1949)
    \item \code{q.nielsen}, a better unbiased transformation of \code{q}
          following Nielsen \emph{et al}. (2003)
}
Each such list \code{q}, \code{r} and \code{q.nielsen} contains several
elements representing different diversity statistics.  Several
quantities are used below:
\itemize{
    \item \code{q.gg}, vector LENGTH of squared frequencies of groups in each
          site
    \item \code{q.bar.0}, scalar, weighted mean of \code{q.gg}
    \item \code{q.unweighted.mean}, mean of \code{q.gg}
    \item \code{alpha.g}, vector LENGTH of reciprocals of \code{q.gg}
    \item \code{alpha.unweighted.mean}, mean of \code{alpha.g}
    \item \code{d.alpha}, reciprocal of \code{q.unweighted.mean}
    \item \code{d.gamma}, reciprocal of sum of grand squared frequencies
    \item \code{d.beta}, \code{d.gamma / d.alpha}
    \item \code{diversity.mat} DIMS, Chao pairwise off-diagonal,
          \code{alpha.g} on-diagonal
    \item \code{divergence.mat} DIMS, \code{diversity.mat} with \code{0}
          on-diagonal
    \item \code{overlap.mat}, \code{1 - divergence.mat}
    \item \code{overlap}, mean overlap, \eqn{\sum{C} / ((G - 1) \sum{q_{gg}})}
    \item \code{divergence}, mean divergence, \code{1 - overlap}
}
The list also contains the original table passed to \code{diversity},
together with several summary quantities:
\itemize{
    \item \code{table}, the table passed in as \code{tab}
    \item \code{num.groups}, the number of rows in \code{tab}
    \item \code{num.sources}, the number of columns in \code{tab}
    \item \code{num.samples}, the sum of all elements in \code{tab}
    \item \code{num.samples.group}, a vector of length \code{num.groups},
          the sum of elements in each row of \code{tab}
    \item \code{num.sources.group}, a vector of length \code{num.groups},
          the number of distinct groups observed in each row of \code{tab}
    \item \code{num.samples.source}, a vector of length \code{num.sources},
          the sum of elements in each column of \code{tab}
    \item \code{num.groups.source}, a vector of length \code{num.groups},
          the number of distinct groups observed in each row of \code{tab}
}
Finally, some additional statistics representing pooled PMI measures
(Scofield \emph{et al}. 2010) are returned:
\itemize{
    \item \code{y.gh}
    \item \code{prop.y.0.gh}
    \item \code{q.0.gh}
}
}
\description{
Given a site x source matrix, calculate a collection of diversity
statistics.  Statistics include PMI (probability of maternal inheritance)
from Grivet \emph{et al}. (2005) and Scofield \emph{et al}. (2010, 2011)
as well as alpha, beta, gamma and delta/omega from classical ecology and
applied to a dispersal context by Scofield \emph{et al}. (2012).
}
\references{
Simpson, E. D. (1949) Measurement of diversity. \emph{Nature} 163:688.

Chao, A., Jost, L., Chiang, S. C., Jiang, Y. H., and Chazdon, R. L. (2008)
A two-stage probabilistic approach to multiple-community similarity
indices.  \emph{Biometrics} 64:1178-1186.

Grivet, D., Smouse, P. E. and Sork, V. L. (2005) A novel approach to an old
problem: tracking dispersed seeds.  \emph{Molecular Ecology} 14:3585-3595.

Nielsen, R., Tarpy, D. R. and Reeve, H. K. (2003) Estimating effective
paternity number in social insects and the effective number of alleles in
a population.  \emph{Molecular Ecology} 12:3157-3164.

Scofield, D. G.,  Sork, V. L. and Smouse, P. E. (2010) Influence of acorn
woodpecker social behaviour on transport of coast live oak
(\emph{Quercus agrifolia}) acorns in a southern California oak savanna.
\emph{Journal of Ecology} 98:561-572.

Scofield, D. G., Alfaro, V. R., Sork, V. L., Grivet, D., Martinez, E.,
Papp, J., Pluess, A. R., Koenig, W. D. and Smouse, P. E. (2011) Foraging
patterns of acorn woodpeckers (\emph{Melanerpes formicivorus}) on valley
oak (\emph{Quercus lobata} N\'{e}e) in two California oak savanna-woodlands.
Oecologia 166:187-196.

Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}

