% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diversity-allele.R
\name{diversitySingleLocus}
\alias{diversitySingleLocus}
\title{Calculate allelic diversity for a single locus}
\usage{
diversitySingleLocus(tab, method = c("r", "q.nielsen", "q"))
}
\arguments{
\item{tab}{Table of site-by-allele counts, with row names being the site
names and column names being the names given to the individual alleles}

\item{method}{Method to use when calculating diversity within the locus.
See \code{\link{diversity}}.}
}
\value{
List containing diversity estimates for the locus
}
\description{
The single argument is, for a single locus, a table of site-by-allele
counts, with row names being the site names, and column names being the
names given to the individual alleles.
}

