% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diversityTests-allele.R
\name{gammaContrastTest}
\alias{gammaContrastTest}
\alias{gammaContrastTest.allele_tables}
\title{Test for difference in gamma diversity between two allele diversity datasets}
\usage{
\method{gammaContrastTest}{allele_tables}(lst.a, lst.b,
  zero.var.adjust = TRUE, n.resample = 10000, method = c("bootstrap",
  "permute"), test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99,
  0.999))
}
\arguments{
\item{lst.a}{Allele diveristy dataset, a list, one entry per locus, of
              site x allele counts. Each table must have the same format.}

\item{lst.b}{Allele diveristy dataset, a list, one entry per locus, of
              site x allele counts. Each table must have the same format.}

\item{\dots}{Additional parameters}
}
\value{
Fill this in
}
\description{
From Sork et al., extensions of those from Scofield et al. 2012 American
Naturalist 180(6) 719-732, http://www.jstor.org/stable/10.1086/668202).
}
\examples{
# For comparing allele diversity between two different samples:

# dat1 <- readGenalex("file-of-genotypes-sample-1.txt")
# dat2 <- readGenalex("file-of-genotypes-sample-2.txt")
# gt1 <- createAlleleTables(dat1)
# gt2 <- createAlleleTables(dat2)
# gamma.contrast <- gammaContrastTest(gt1, gt2)
}

