% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{gammaContrastTest3}
\alias{gammaContrastTest3}
\alias{gammaContrastTest3.divtable}
\title{Test for difference in gamma diversity among three datasets}
\usage{
\method{gammaContrastTest3}{divtable}(tab.a, tab.b, tab.c,
  zero.var.adjust = TRUE, n.resample = 10000, method = c("bootstrap",
  "permute"), test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99,
  0.999), ...)
}
\arguments{
\item{tab.a}{First site-by-source table}

\item{tab.b}{Second site-by-source table}

\item{tab.c}{Third site-by-source table}

\item{zero.var.adjust}{Logical, whether to adjust zero-variance groups
with minimum value, see \code{\link{BLAHBLAH}}}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}
}
\value{
Class \code{diversity_test} object with the result of the test
}
\description{
Test for difference in gamma diversity among three datasets
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{gammaContrastTest}}, \code{\link{alphaContrastTest3}}
}

