% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dispersalDiversity-data.R
\docType{data}
\name{granaries_2002_Qlob}
\alias{granaries_2002_Qlob}
\title{Assignments of acorns harvested from acorn woodpecker granaries in 2002,
to seed source trees}
\format{Matrix with 21 rows and 50 variables.}
\source{
Dataset originates from the lab of Victoria L. Sork
          \email{vlsork@ucla.edu} and is used with permission.
}
\description{
A dataset containing seed source tree assignments for acorns of
valley oak (\emph{Quercus lobata}), harvested in 2002 from acorn
woodpecker (\emph{Melanerpes formicivorus}) granaries. Data were collected
in oak savanna habitat in central California.  Acorn woodpeckers store
acorns in central granaries, and different woodpecker social groups
maintain different granaries.
}
\details{
The dataset is stored as a matrix, with the variables are as follows:
\itemize{
    \item Granary_Year, the row names, indicating the granary identifier
        and year of harvest
    \item Source, the column names, indicating the identified source of
        harvested acorns.  Where the source name is prefixed with "S",
        the source tree was identified among the adult trees genotyped at
        the study site.  Where the prefix is "U", the source tree was not
        genotyped and thus not located, but could still be identified as
        unique among the harvested acorns.
}
Each element of the matrix represents a count of the number of acorns
harvested from that granary in 2002 that were assigned to that seed
source tree.

Some, but not all, granaries and seed sources are common between this
dataset and \code{granaries_2004_Qlob}, also available in this package.
See Thompson \emph{et al}. (2014) for more study details.
}
\author{
Douglas G. Scofield \email{douglasgscofield@gmail.com}
}
\references{
Thompson, P. G., Smouse, P. E., Scofield, D. G. and Sork, V. L. (2014)
What seeds tell us about birds: a multi-year analysis of acorn woodpecker
foraging movements.  \emph{Movement Ecology} 2:12.
\url{http://www.movementecologyjournal.com/content/2/1/12}
}
\keyword{datasets}

