% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{pairwiseMeanTest}
\alias{pairwiseMeanTest}
\alias{pairwiseMeanTest.divtable}
\title{Test for structure in pairwise divergence between sites}
\usage{
\method{pairwiseMeanTest}{divtable}(tab, n.iter = 10000, method = c("r",
  "q", "q.nielsen"), statistic = c("divergence", "overlap"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999),
  ...)
}
\description{
Test for structure in pairwise divergence between sites
}

