% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{plot.alpha_contrast_test3}
\alias{plot.alpha_contrast_test3}
\title{Plot result of \code{alphaContrastTest3}}
\usage{
\method{plot}{alpha_contrast_test3}(result, add.analytic = FALSE,
  compress.x = TRUE, ...)
}
\description{
Plot result of \code{alphaContrastTest3}
}

