% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{plot.alpha_test}
\alias{plot.alpha_test}
\title{Plot result of \code{alphaDiversityTest}}
\usage{
\method{plot}{alpha_test}(result, add.analytic = FALSE, compress.x = TRUE,
  ...)
}
\description{
Plot result of \code{alphaDiversityTest}
}

