% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gammaAccum.R
\name{plot.gamma_accum}
\alias{plot.gamma_accum}
\title{Plot the gamma accumulation result from \code{gammaAccum}}
\usage{
\method{plot}{gamma_accum}(g, xmax = length(g$simple.results$mns),
  ymax = max(g$simple.results$mns + g$simple.results$SE), obs.omega = NULL,
  omega.pos = c(0.15, 0.97), xlab = "Number of groups",
  ylab = expression("Accumulated  " * gamma * "  diversity"), pch = 19,
  pch.gamma = 24, col = "black", bty = "L", lwd = 1.5, bg = "white",
  lty = 2, ...)
}
\arguments{
\item{g}{Object of class \code{'gamma_accum'} returned by
\code{\link{gammaAccum}}}

\item{xmax}{Maximum extent of the X-axis in the plot}

\item{ymax}{Maximum extent of the Y-axis in the plot}

\item{obs.omega}{If provided, the observed omega value for the
data underlying the gamma accumulation curve; this value is added to
the plot}

\item{omega.pos}{If \code{obs.omega} is given, the location at
which it is plotted, relative to \code{xmax} and \code{ymax},
in a two-element vector.  1 is added to each dimension.}

\item{xlab,ylab,col,bty,lwd,bg}{Options passed to \code{\link{plot}},
\code{\link{lines}} and/or \code{\link{points}}}

\item{pch}{Plotting character used for mean gamma values accumulated,
see \code{?pch}}

\item{pch.gamma}{Plotting character used for total gamma value,
\code{?pch}}

\item{\dots}{Additional options passed to \code{\link{plot}},
\code{\link{lines}} and/or \code{\link{points}}}
}
\value{
Nothing returned
}
\description{
For an example of its use see Figure 4D-F of Scofield et al. (2012).
}
\examples{
##TODO: find a tab
#rga.result = gammaAccum(tab)
#plot(rga.result)
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{gammaAccum}}
}

