% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{plot.gamma_contrast_test}
\alias{plot.gamma_contrast_test}
\title{Plot result of \code{gammaContrastTest}}
\usage{
\method{plot}{gamma_contrast_test}(result, add.analytic = FALSE,
  compress.x = TRUE, ...)
}
\description{
Plot result of \code{gammaContrastTest}
}

