% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{plot.pairwise_mean_test}
\alias{plot.pairwise_mean_test}
\title{Plot the result of \code{'pairwiseMeanTest'}}
\usage{
\method{plot}{pairwise_mean_test}(result, ...)
}
\description{
Plot the result of \code{'pairwiseMeanTest'}
}

