% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gammaAccum.R
\name{plotGammaAccum}
\alias{plotGammaAccum}
\title{Plot the gamma accumulation result from \code{runGammaAccum}}
\usage{
plotGammaAccum(gamma.accum,
  plot.xmax = length(gamma.accum$simple.results$mns),
  plot.ymax = max(gamma.accum$simple.results$mns +
  gamma.accum$simple.results$SE), obs.omega = NULL,
  omega.positions = c(0.15, 0.97), xlab = "Number of seed pools",
  ylab = expression("Accumulated  " * gamma * "  diversity"), pch = 19,
  pch.gamma = 24, col = "black", bty = "L", lwd = 1.5, bg = "white",
  lty = 2, ...)
}
\arguments{
\item{gamma.accum}{Result from \code{\link{runGammaAccum}}}

\item{plot.xmax}{Maximum extent of the X-axis in the plot}

\item{plot.ymax}{Maximum extent of the Y-axis in the plot}

\item{obs.omega}{If provided, the observed omega value for the
data underlying the gamma accumulation curve; this value is added to
the plot}

\item{omega.positions}{If \code{obs.omega} is given, the location at
which it is plotted, relative to \code{plot.xmax} and \code{plot.ymax},
in a two-element vector.  1 is added to each dimension.}

\item{xlab,ylab,col,bty,lwd,bg}{Options passed to \code{\link{plot}},
\code{\link{lines}} and/or \code{\link{points}}}

\item{pch}{Plotting character used for mean gamma values accumulated,
see \code{?pch}}

\item{pch.gamma}{Plotting character used for total gamma value,
\code{?pch}}

\item{\dots}{Additional options passed to \code{\link{plot}},
\code{\link{lines}} and/or \code{\link{points}}}
}
\value{
Nothing returned
}
\description{
For an example of its use see Figure 4D-F of Scofield et al. (2012).
}
\seealso{
\code{\link{runGammaAccum}}
}

