% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPairwiseMatrix.R
\name{plotPairwiseMatrix}
\alias{plotPairwiseMatrix}
\title{Plot pairwise divergence or overlap as calculated by \code{diversity}}
\usage{
plotPairwiseMatrix(pairwise.mat, statistic = c("divergence", "overlap"),
  pairwise.mean = NULL, mean.position = c(0.45, 0.7), bty = "L",
  aspect = "iso", col.regions = function(x) gray(c(1, seq(0.9, 0.6,
  length.out = (x - 2)), 0)), colorkey = list(labels = list(cex = 1)),
  at = c(0, 0.01, seq(0.2, 0.8, 0.2), 0.99, 1), scales = list(draw = FALSE,
  tck = 0, cex = 0.7, x = list(rot = 90)), axis.label = "Species Pool",
  xlab = list(axis.label, cex = 1), ylab = list(axis.label, cex = 1), ...)
}
\arguments{
\item{pairwise.mat}{Pairwise divergence or overlap matrix as found at,
e.g., \code{diversity()$q$diversity.mat}.  Currently, prior to plotting
this matrix has its diagonal and upper triangle zeroed, and is then
rotated prior to passing to \code{\link{lattice::levelplot}}}

\item{statistic}{\code{"divergence"} or \code{"overlap"}, for which
statistic is being presented}

\item{pairwise.mean}{Mean pairwise divergence or overlap as found at,
e.g., \code{diversity()$q$divergence}.  If provided, this is added to
the plot in the upper triangle, rounded to three digits, with
positions specified by \code{mean.pos}.  The value is plotted
together with '\eqn{\bar{\delta} = }' if \code{statistic} is
\code{"divergence"}, and '\eqn{\bar{\omega} = }' if \code{statistic} is
\code{"overlap"}.}

\item{bty,aspect,col.regions,colorkey,at,scales,xlab,ylab}{Additional
plot options passed to \code{\link{lattice::levelplot}}}

\item{axis.label}{Label used for the X and Y axes, the X and Y axis
labels can be changed with \code{xlab} and \code{ylab}}

\item{mean.pos}{If \code{pairwise.mean} is given, the relative
X and Y positions within the panel at which the value is plotted, in
a two-element vector.  \code{adj = c(0, 0)} is used when plotting the
value.}
}
\value{
The lattice plot object is returned invisibly
}
\description{
Plot pairwise values using \code{levelplot} from the \code{lattice}
package.  An example of its use is in Figure 4A-C of Scofield et al.
(2012).
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{diversity}}, \code{\link{lattice::levelplot}}
}

