% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-allele_divtables.R
\name{print.allele_diversity_test}
\alias{print.allele_diversity_test}
\title{Print the result of a allele diversity test (alpha or gamma)}
\usage{
\method{print}{allele_diversity_test}(x, digits = getOption("digits"),
  include.loci = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{allele_diversity_test}, returned by
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest}} and
\code{\link{gammaContrastTest}} when given an argument of class
\code{\link{allele_divtables}}}

\item{digits}{Number of significant digits to use when printing
numeric values.  Defaults to the value of the \code{"digits"} option.}

\item{include.loci}{If \code{TRUE}, print test results for each
individual locus.  Defaults to \code{FALSE}.}
}
\value{
\code{x}, returned invisibly
}
\description{
Print an object of class \code{allele_diversity_test}, the result of
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest}} and
\code{\link{gammaContrastTest}} when given an argument of class
\code{\link{allele_divtables}}.
}
\details{
The result of the test printed is the summed effect across all loci;
see Sork \emph{et al}. (2015) for details.  If \code{include.loci} is
\code{TRUE}, then this is followed by the test result for each locus
individually.
}
\examples{
## library(readGenalex)  # already loaded as a prerequisite
data(Qagr_pericarp_genotypes)  # from readGenalex
gt <- createAlleleTables(Qagr_pericarp_genotypes)
alpha.test <- alphaDiversityTest(gt)
print(alpha.test)
}
\references{
Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (In press)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).  \emph{Evolutionary Ecology}.
}
\seealso{
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest}}, \code{\link{gammaContrastTest}}
}

