% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{print.alpha_contrast_test3}
\alias{print.alpha_contrast_test3}
\title{Print the result of \code{alphaContrastTest3}}
\usage{
\method{print}{alpha_contrast_test3}(result, ...)
}
\description{
Print the result of \code{alphaContrastTest3}
}

