% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{print.diversity_test}
\alias{print.diversity_test}
\title{Print the result of a diversity test (alpha or gamma)}
\usage{
\method{print}{diversity_test}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object of class \code{diversity_test}, returned by
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest}},
\code{\link{alphaContrastTest3}}, \code{\link{gammaContrastTest}} and
\code{\link{gammaContrastTest3}} when given an argument of class
\code{\link{allele_divtables}}.}

\item{digits}{Number of significant digits to use when printing
numeric values}
}
\value{
\code{x}, returned invisibly
}
\description{
Print an object of class \code{diversity_test}, the result of
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest}},
\code{\link{alphaContrastTest3}}, \code{\link{gammaContrastTest}} and
\code{\link{gammaContrastTest3}} when given an argument of class
\code{\link{divtable}}.
}
\seealso{
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest}}, \code{\link{alphaContrastTest3}}, \code{\link{gammaContrastTest}}, \code{\link{gammaContrastTest3}}
}

