% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{print.gamma_contrast_test}
\alias{print.gamma_contrast_test}
\title{Print the result of \code{gammaContrastTest}}
\usage{
\method{print}{gamma_contrast_test}(result, ...)
}
\description{
Print the result of \code{gammaContrastTest}
}

