% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{print.gamma_contrast_test3}
\alias{print.gamma_contrast_test3}
\title{Print the result of \code{gammaContrastTest3}}
\usage{
\method{print}{gamma_contrast_test3}(result, ...)
}
\description{
Print the result of \code{gammaContrastTest3}
}

