% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gammaAccum.R
\name{runGammaAccum}
\alias{runGammaAccum}
\title{Perform gamma diversity accumulation on site-by-source data}
\usage{
runGammaAccum(tab, accum.method = c("random", "proximity"),
  resample.method = c("permute", "bootstrap"), gamma.method = c("r",
  "q.nielsen", "q"), distance.file = NULL, ...)
}
\arguments{
\item{tab}{Site-by-source table, as passed to diversity()}

\item{accum.method}{\code{"random"} (default) or \code{"proximity"}.
Method for accumulating sites.  If \code{"random"}, then the next site is
chosen at random. If \code{"proximity"}, then the next site chosen is the
next closest site, and \code{distance.file} must be supplied.}

\item{resample.method}{\code{"permute"} (default) or \code{"bootstrap"}.
Whether to resample sites without replacement (\code{"permute"}) or with
replacement (\code{"bootstrap"}).}

\item{gamma.method}{Calculate gamma using \code{"r"} (default),
\code{"q"} or \code{"q.nielsen"} method (see paper).}

\item{distance.file}{A file or data.frame containing three columns of data,
with the header/column names being \code{"pool"}, \code{"X"}, and \code{"Y"},
containing the spatial locations of the seed pools named in the row names of
tab.  Only used if the \code{accum.method=="proximity"}.}

\item{\dots}{Additional parameters passed to \code{runGammaAccumSimple}}
}
\description{
Perform a gamma diversity accumulation on the site-by-source data in tab.
Several arguments control the method of accumulation and value of gamma
calculated.  Only the defaults have been tested; the others were developed
while exploring the data and must be considered experimental.  The result is
returned in a list, which may be passed to \code{plotGammaAccum} to plot the
result.
}
\seealso{
\code{\link{plotGammaAccum}}, \code{\link{runGammaAccumSimple}}
}

