% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_divtables-class.R
\name{allele_divtables-class}
\alias{allele_divtables}
\alias{allele_divtables-class}
\title{List of divtables holding allele diversity data}
\description{
An object of class \code{allele_divtables} is a list of
\code{\link{divtable}} objects, each representing sites-by-allele counts
data for a single genetic locus.  Row and column names are
the site names and individual alleles, respectively.
This is the basic data object for analysis of genetic data using the
\code{\link{dispersalDiversity}} package.  It is accepted by the
function \code{\link{diversity}}, which generates descriptive statistics,
and the functions \code{\link{alphaDiversityTest}},
\code{\link{gammaContrastTest}} and others that test for differences
in the structure of genetic diversity.
}
\details{
Assembling an \code{allele_divtables} by hand is labourious.  More
typically, genotype data will be in a class \code{\link{genalex}} object
and then converted to \code{\link{allele_divtables}} using
\code{\link{createAlleleTables}}.  The \code{\link{createAlleleTables}}
and \code{\link{as.allele_divtables}} functions will attempt to convert
non-\code{\link{genalex}} objects using to \code{\link{genalex}} format
using \code{\link[readGenalex]{as.genalex}}, but this may not be successful.
}
\examples{

## One possible way to plot an \\code{allele_divtables} object, plotting
## each locus as a separate divtable in a two-column format

data(Qagr_pericarp_genotypes)
pal <- createAlleleTables(Qagr_pericarp_genotypes)
par(mfrow = c(round(length(pal) / 2), 2))
lapply(names(pal), function(n) plot(pal[[n]], main = n, l2 = NULL, las = 2))

}

