% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversityTests.R, R/diversityTests.R
\name{alphaContrastTest}
\alias{alphaContrastTest}
\alias{alphaContrastTest.allele_tables}
\title{Test for difference in alpha diversity between two allele diversity datasets}
\usage{
\method{alphaContrastTest}{allele_tables}(lst.a, lst.b,
  zero.var.adjust = TRUE, n.resample = 10000, method = c("bootstrap",
  "permute"), test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99,
  0.999))

alphaContrastTest(x, ...)
}
\arguments{
\item{lst.a}{Allele diveristy dataset, a list, one entry per locus, of
              site x allele counts. Each table must have the same format.}

\item{lst.b}{Allele diveristy dataset, a list, one entry per locus, of
              site x allele counts. Each table must have the same format.}

\item{zero.var.adjust}{Logical, whether to adjust zero-variance groups
with minimum value, see \code{\link{BLAHBLAH}}}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}

\item{\dots}{Additional parameters}

\item{tab.a}{First ite x soure table}

\item{tab.b}{Second site x soure table}
}
\value{
Fill this in

An \code{'htest_boot'} object with the result of the test. The
test also prints the result, so perhaps I should modify it to only
return the htest_boot object?  Or should it return a different object?
}
\description{
From Sork et al., extensions of those from Scofield et al. 2012 American
Naturalist 180(6) 719-732, http://www.jstor.org/stable/10.1086/668202).

Test for differences in alpha diversity between two data sets
}
\examples{
# For comparing allele diversity between two different samples:

# dat1 <- readGenalex("file-of-genotypes-sample-1.txt")
# dat2 <- readGenalex("file-of-genotypes-sample-2.txt")
# gt1 <- createAlleleTables(dat1)
# gt2 <- createAlleleTables(dat2)
# alpha.contrast <- alphaContrastTest(gt1, gt2)
}
\author{
Douglas G. Scofield
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{}}

\code{\link{alphaContrastTest.3}}
}

