% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-allele_divtables.R, R/diversityTests-divtable.R
\name{alphaContrastTest.allele_divtables}
\alias{alphaContrastTest}
\alias{alphaContrastTest.allele_divtables}
\alias{alphaContrastTest.default}
\alias{alphaContrastTest.divtable}
\title{Test for differences in alpha diversity between two data sets}
\usage{
\method{alphaContrastTest}{allele_divtables}(adt.a, adt.b,
  zero.div.adjust = TRUE, n.resample = 10000, method = c("bootstrap",
  "permute"), test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99,
  0.999))

\method{alphaContrastTest}{divtable}(tab.a, tab.b, zero.div.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999))

\method{alphaContrastTest}{default}(tab.a, tab.b, ...)
}
\arguments{
\item{adt.a}{Allelic diversity dataset of class
\code{\link{allele_divtables}}}

\item{adt.b}{Allelic diversity dataset of class
\code{\link{allele_divtables}}}

\item{zero.div.adjust}{Logical, if \code{TRUE} (the default), then groups
with 0 within-group diversity are assigned a minimum diversity which is
half the empirical diversity possible given the group size}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}

\item{tab.a}{Site-by-source table of class \code{\link{divtable}}}

\item{tab.b}{Site-by-source table of class \code{\link{divtable}}}
}
\value{
An \code{diversity_test} object with the result of the test
}
\description{
The null hypothesis for this tests is that there is no difference in alpha
diversity between the two sets of datasets sites represented in \code{tab.a}
and \code{tab.b}, or \code{adt.a} and \code{adt.b}.  The method for the
initial class \code{\link{divtable}} version of this was described in
Scofield et al. (2012), while that for the allelic extension as class
\code{link{allele_divtables}} was described in Sork \emph{et al}.
(In press).
}
\note{
\code{adt.a} and \code{adt.b} must contain the same loci
}
\examples{
## Comparing alpha diversity between two different sites:
##
data(granaries_2002_Qlob)
data(granaries_2004_Qlob)
par(mfcol = c(2, 1))
membershipPlot(granaries_2002_Qlob)
membershipPlot(granaries_2004_Qlob)
alphaContrastTest(granaries_2002_Qlob, granaries_2004_Qlob)

## Comparing allele diversity between two different samples:
##
# dat1 <- readGenalex("file-of-genotypes-sample-1.txt")
# dat2 <- readGenalex("file-of-genotypes-sample-2.txt")
# gt1 <- createAlleleTables(dat1)
# gt2 <- createAlleleTables(dat2)
# alpha.contrast <- alphaContrastTest(gt1, gt2)
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.

Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (In press)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).  \emph{Evolutionary Ecology}.
}
\seealso{
\code{\link{alphaDiversityTest}}, \code{\link{alphaContrastTest3}}
}

