% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-allele_divtables.R, R/diversityTests-divtable.R
\name{alphaDiversityTest}
\alias{alphaDiversityTest}
\alias{alphaDiversityTest.allele_divtables}
\alias{alphaDiversityTest.divtable}
\title{Test for alpha diversity difference between sites}
\usage{
\method{alphaDiversityTest}{allele_divtables}(lst, zero.var.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999),
  ...)

\method{alphaDiversityTest}{divtable}(tab, zero.var.adjust = TRUE,
  n.resample = 10000, method = c("bootstrap", "permute"),
  test.quantiles = c(0.001, 0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99, 0.999),
  ...)
}
\arguments{
\item{lst}{Object of class \code{'allele_divtables'}, as produced by the
function \code{\link{createAlleleTables}}.  A list, one entry per locus,
of site x allele counts.  Each table must have the same format.}

\item{zero.var.adjust}{Logical, whether to adjust zero-variance groups
with minimum value, see \code{\link{BLAHBLAH}}}

\item{n.resample}{Number of iterations for creation of the null distribution}

\item{method}{\code{"bootstrap"} or \code{"permute"}, whether to create null
distribution iterations with (\code{"bootstrap"}) or without
(\code{"permute"}) replacement}

\item{tab}{Site x source table}

\item{\dots}{Additional parameters}
}
\value{
Fill this in

An \code{'htest_boot'} object with the result of the test. The
test also prints the result, so perhaps I should modify it to only
return the htest_boot object?  Or should it return a different object?
}
\description{
From Sork et al., extensions of those from Scofield et al. 2012 American
Naturalist 180(6) 719-732, http://www.jstor.org/stable/10.1086/668202).

Test for differences in alpha diversity among sites within a single data set
}
\examples{
## For comparing allele diversity between sites in the same sample
#
# dat <- readGenalex("file-of-genotypes-sample-1.txt")
# gt <- createAlleleTables(dat)
# alpha.test <- alphaDiversityTest(gt)
}
\author{
Douglas G. Scofield
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.

Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (Submitted)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).

Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}
\seealso{
\code{\link{alphaDiversityTest}}, \code{\link{diversity}}

\code{\link{alphaContrastTest}}
}

