% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divtable-class.R
\name{as.divtable}
\alias{as.divtable}
\alias{as.divtable.data.frame}
\alias{as.divtable.default}
\alias{as.divtable.matrix}
\alias{as.divtable.table}
\alias{as.divtable.xtabs}
\title{Convert to class divtable}
\usage{
\method{as.divtable}{table}(x, ...)

\method{as.divtable}{matrix}(x, ...)

\method{as.divtable}{data.frame}(x, ...)

\method{as.divtable}{xtabs}(x, ...)

\method{as.divtable}{default}(x, ...)
}
\description{
Convert to class divtable
}

