% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allele_divtables-class.R
\name{createAlleleTables}
\alias{createAlleleTables}
\alias{createAlleleTables.genalex}
\title{Calculate an \code{\link{allele_divtables}} object from a class \code{genalex} object}
\usage{
\method{createAlleleTables}{genalex}(x, new.ploidy = 2,
  collapse.alleles = TRUE, exclude = c(NA, "0"), quiet = FALSE)
}
\description{
S3 method to convert an object of class \code{genalex} to an object of
class \code{\link{allele_divtables}}, a list of \code{\link{divtable}}
objects representing site-specific allele counts.
Each entry of the list is, for each locus, a
table of site-by-allele counts, with row names being the site names and
column names being the names given to the individual alleles.  This is a
generic so that other methods might be written to convert other genetic
formats.
}
\examples{
# The workflow to calculate basic allelic diversity statistics:
#
# dat <- readGenalex("GenAlEx-format-file-of-genotypes.txt")
# gt <- createAlleleTables(dat)
}

