% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversity.R
\name{createAlleleTables}
\alias{createAlleleTables}
\alias{createAlleleTables.genalex}
\title{Convert class \code{'genalex'} object to a list of allele count tables}
\usage{
\method{createAlleleTables}{genalex}(dat, new.ploidy = 2,
  collapse.alleles = TRUE, exclude = c(NA, "0"), quiet = FALSE)
}
\description{
S3 method to convert an object of class \code{'genalex'} to a list of
allele count tables.  This is a generic so that other methods might be
written to convert other genetic formats.
}
\examples{
# The workflow to calculate basic allelic diversity statistics:
#
# dat <- readGenalex("GenAlEx-format-file-of-genotypes.txt")
# gt <- createAlleleTables(dat)
}

