% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity-allele_divtables.R
\name{diversityMultilocus}
\alias{diversityMultilocus}
\alias{diversityMultilocus.allele_divtables}
\title{Calculate diversity across a set of allele tables}
\usage{
\method{diversityMultilocus}{allele_divtables}(x, ploidy = 2,
  method = c("r", "q.nielsen", "q"), ...)
}
\arguments{
\item{x}{Allele diversity data contained in an object of class
\code{\link{allele_divtables}}}

\item{ploidy}{Ploidy of underlying genotypes}

\item{method}{Method to use for calculating diversity statistics.
See \code{\link{diversity}} and Scofield \emph{et al}. (2012),
Sort \emph{et al}. (in press).}
}
\value{
A list of class \code{multilocus_diversity}.  Elements of the
list are calculated from values returned by
\code{\link{diversitySingleLocus}} and include:
\itemize{
    \item \code{q.gg}, matrix of \code{allele.alpha.g} values across sites
          (rows), for each locus (columns)
    \item \code{alpha.gk}, matrix of \code{allele.alpha.g} values across
          sites (rows), for each locus (columns) (\emph{are these simply
          reciprocals of q.gg?})
    \item \code{q.k}, vector of mean \code{q.gg} values across loci, the
          mean values of the rows of \code{q.gg}
    \item \code{alpha.k}, vector of mean \code{allele.alpha.g} values
          across loci, the mean values of the rows of \code{alpha.gk}
    \item \code{q.bar.weighted.a}, vector of \code{allele.q.bar.weighted}
          for each locus
    \item \emph{\code{alpha.bar.weighted.a}, vector of
          \code{allele.alpha.bar.weighted} for each locus ????}
    \item \emph{\code{overlap.a}, vector of \code{allele.overlap} for each locus ????}
    \item \emph{\code{divergence.a}, vector of \code{allele.divergence} for each locus ????}
    \item \code{Q.0.a}, vector of \code{allele.Q.0} values for each locus
    \item \code{q.bar.weighted}, mean of \code{q.bar.weighted.a}
    \item \code{alpha.bar.weighted}, reciprocal of \code{q.bar.weighted}
    \item \emph{\code{Q.0}, mean of \code{Q.0.a} ????}
    \item \code{gamma}, reciprocal of mean of \code{Q.0.a}, \emph{Q.0 ????}
    \item \code{beta}, \code{gamma} divided by \code{alpha.bar.weighted}
    \item \code{G}, number of loci
    \item \code{N}, allele count for in each locus
    \item \code{K}, number of sites
    \item \code{alpha.scaled}, \code{alpha.bar.weighted} scaled to fall
          in the interval [0, 1]
    \item \code{gamma.scaled}, \code{gamma} scaled to fall in the interval
          [0, 1]
    \item \code{beta.scaled}, \code{beta} scaled to fall in the interval
          [0, 1]
    \item \code{q.gh}, site-by-site matrix of mean of \code{q.gh} values
          across all loci
    \item \code{overlap}, sum of \code{q.gh} divided by \code{K} - 1 times
          the sum of \code{q.k}
    \item \code{divergence}, 1 - \code{overlap}
    \item \code{overlap.mean}, mean of \code{allele.divergence} values
          for each locus (\emph{\code{overlap.a} ???})
    \item \code{divergence.mean}, mean of \code{allele.overlap} values
          for each locus (\emph{\code{divergence.a} ???})
}

These items were left out for no apparent reason or are confusing:
\itemize{
    \item \code{Q.0}
    \item \code{alpha.bar.weighted.a}
    \item \code{G}
    \item \code{N}
    \item \code{K}
    \item \code{overlap.a}
    \item \code{divergence.a}
}

  ans <- list(q.gg               = q.gg,
              alpha.gk           = alpha.gk,
              q.k                = q.k,
              alpha.k            = alpha.k,
              q.bar.weighted.a   = q.bar.weighted.a,
              Q.0.a              = Q.0.a,
              q.bar.weighted     = q.bar.weighted,
              alpha.bar.weighted = alpha.bar.weighted,
              gamma              = gamma,
              beta               = beta,
              G                  = G,
              N                  = N,
              K                  = K,
              alpha.scaled       = alpha.scaled,
              gamma.scaled       = gamma.scaled,
              beta.scaled        = beta.scaled,
              q.gh               = q.gh,
              overlap            = overlap,
              divergence         = divergence,
              overlap.mean       = overlap.mean,
              divergence.mean    = divergence.mean)
}
\description{
Calculate diversity for a set of loci encoded in an object of class
\code{'allele_tables'} as created with \code{\link{createAlleleTables}},
and diversity for each locus is calculated separately
using \code{\link{diversitySingleLocus}}.
}
\examples{
## Workflow to calculate basic allelic diversity statistics:
library(readGenalex)
data(Qagr_pericarp_genotypes)
gt <- createAlleleTables(Qagr_pericarp_genotypes)
div <- diversityMultilocus(gt)
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.

Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (In press)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).  \emph{Evolutionary Ecology}.
}

