% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleleDiversity.R
\name{diversityMultilocus}
\alias{diversityMultilocus}
\alias{diversityMultilocus.allele_tables}
\title{Calculate diversity across a set of allele tables}
\usage{
\method{diversityMultilocus}{allele_tables}(lst, method = c("r", "q.nielsen",
  "q"))
}
\description{
Calculate diversity across a set of allele tables
}
\examples{
# The workflow to calculate basic allelic diversity statistics:
#
# dat <- readGenalex("GenAlEx-format-file-of-genotypes.txt")
# gt <- createAlleleTables(dat)
# div <- diversityMultilocus(gt)
}

