% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity-allele_divtables.R
\name{diversityMultilocus}
\alias{diversityMultilocus}
\alias{diversityMultilocus.allele_divtables}
\title{Calculate diversity across a set of allele tables}
\usage{
\method{diversityMultilocus}{allele_divtables}(x, ploidy = 2,
  method = c("r", "q.nielsen", "q"), ...)
}
\description{
Calculate diversity for a set of loci encoded in an object of class
\code{'allele_tables'} as created with \code{\link{createAlleleTables}},
and diversity for each locus is calculated separately
using \code{\link{diversitySingleLocus}}.
}
\examples{
## Workflow to calculate basic allelic diversity statistics:
library(readGenalex)
data(Qagr_pericarp_genotypes)
gt <- createAlleleTables(Qagr_pericarp_genotypes)
div <- diversityMultilocus(gt)
}

