% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gammaAccum-allele_divtables.R, R/gammaAccum-divtable.R
\name{gammaAccum.allele_divtables}
\alias{gammaAccum}
\alias{gammaAccum.allele_divtables}
\alias{gammaAccum.divtable}
\title{Perform gamma diversity accumulation on \code{divtable} or \code{allele_divtables} objects}
\usage{
\method{gammaAccum}{allele_divtables}(adt, resample.method = c("permute",
  "bootstrap"), n.resample = 1000, gamma.method = c("q", "r", "q.nielsen"),
  ...)

\method{gammaAccum}{divtable}(tab, resample.method = c("permute",
  "bootstrap"), n.resample = 1000, gamma.method = c("q", "r", "q.nielsen"),
  ...)
}
\arguments{
\item{adt}{Allele diversity dataset of class
\code{\link{allele_divtables}}}

\item{resample.method}{\code{"permute"} (default) or \code{"bootstrap"}.
Whether to resample sites without replacement (\code{"permute"}) or with
replacement (\code{"bootstrap"}).}

\item{n.resample}{Number of resamples for accumulation curve}

\item{gamma.method}{Calculate gamma using \code{"q"} (default for class
\code{\link{divtable}}), \code{"r"} (default for class
\code{\link{allele_divtables}}) or \code{"q.nielsen"} method}

\item{tab}{Site-by-source table of class \code{\link{divtable}}}

\item{\dots}{Additional parameters, currently unused}
}
\description{
Perform a gamma diversity accumulation on site-by-source data in \code{tab},
an object of class \code{\link{divtable}}, or a set of site-by-alleles count
data for several loci, an object of classs \code{\link{allele_divtables}}.
Several arguments control the method of accumulation and value of gamma
calculated.  Only the defaults have been tested; the others were developed
while exploring the data and must be considered experimental.  The result is
returned in a list, which may be passed to \code{plotGammaAccum} to plot the
result.
}
\note{
Only the defaults have been tested; other options were developed while
exploring the data and must be considered experimental.
}
\examples{
## generate random divtable
n.sites <- 5
n.sources <- 10
n.samples <- 160
set.seed(75333)
t <- data.frame(site = sample(n.sites, n.samples, replace = TRUE),
                source = round(runif(n.samples) * n.sources + 0.5))
tab <- as.divtable(table(t))
## generate gamma accumulation results
rga.result = gammaAccum(tab)

## generate random allele_divtables
##
##
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.

Sork, V. L., Smouse, P. E., Grivet, D. and Scofield, D. G. (In press)
Impact of asymmetric male and female gamete dispersal on allelic
diversity and spatial genetic structure in valley oak
(\emph{Quercus lobata} N\'{e}e).  \emph{Evolutionary Ecology}.
}
\seealso{
\code{\link{plot.gamma_accum}}, \code{\link{gammaAccumSimple}}
}

