% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/membershipPlot.R
\name{membershipPlot}
\alias{membershipPlot}
\alias{membershipPlot.default}
\alias{membershipPlot.divtable}
\title{Plot a series of bars or pies indicating membership in groups}
\usage{
\method{membershipPlot}{default}(tab, ...)

\method{membershipPlot}{divtable}(tab, method = c("bar", "pie"),
  fill.method = c("color", "bw", "colour"), fill.palette = "Dark2",
  distinguish.multiton = FALSE, xlab = "Site", ylab = "Membership",
  las = 1, x.mar.width = ifelse(las >= 2, 4, 3), cex = 0.7,
  header.cex = 0.9, cex.names = 0.85, l1 = NULL, l2 = NULL,
  pdf.file = NULL, postscript.file = NULL, file.dim = c(5.25, 2), ...)
}
\arguments{
\item{tab}{Object of class \code{\link{divtable}} of counts of sites
(rows) by group membership (columns), or any similar object
(\code{table}, \code{matrix}, \code{data.frame}, \code{xtabs}) that
can be converted with \code{\link{as.divtable}}.  This argument is
standardized by site so values may be counts or proportions.}

\item{method}{\code{"bar"} (the default) or \code{"pie"}, for the form
of plot produced.  Bar plots have received considerably more attention
than pie plots.}

\item{fill.method}{\code{"color"} (the default, with synonym
\code{"colour"}, or \code{"bw"}, for the method of colouring different
groups within the plots.  For \code{"color"}, no more than eight separate
colours are chosen; if there are more than eight groups, the
colours for lower-frequency groups are modified by hatching lines of
varying angles and densities.  For \code{"bw"}, fewer grades of grey are
used, together with black-and-white patterns and hatching.

For \code{"color"}, if the package \code{\link{RColorBrewer}} is
available, colours will be chosen from its \code{"Dark2"} palette.
This may be modified with the \code{fill.pallete} option.  The aesthetics
of color choice are important for distinguishing among groups in
membership plots.  If \code{\link{RColorBrewer}} is not available, colours
are chosen using \code{\link{rainbow}}.}

\item{fill.palette}{For \code{fill.method = "color"}, if the
package \code{\link{RColorBrewer}} is available, use this palette to
choose the base colours to use}

\item{distinguish.multiton}{For \code{fill.method = "color"}, whether
to distinguish multiton groups (see Description) with grey rather than
the default white colour.}

\item{xlab,ylab,las,cex}{Plot parameters (see \code{\link{par}}}

\item{x.mar.width}{Width of the X-margin, in a bar plot}

\item{header.cex}{Character expansion factor for plot header}

\item{cex.names}{Character expansion factor for the names plotted below
the bars in a bar plot}

\item{l1,l2}{First and second lines of annotation to appear above
bars in a bar plot}

\item{pdf.file,postscript.file}{Filename for output of PDF file, using
\code{\link{pdf}}, or PostScript file, using \code{\link{postscript}}.
file.  Only one can be set.  Either file is produced using the options
\code{onefile = FALSE, paper = "special"}.}

\item{file.dim}{Dimensions of the plot, in inches}
}
\value{
No value is returned
}
\description{
Singleton groups, those with just one member, appear as white regions
within their respective sites.  Multiton groups, with more than one
member but all members appearing within a single site, also appear as
white regions unless \code{fill.method = "color"} with
\code{distinguish.multiton = TRUE}, then they are grey.  Groups with
members in more than one site are given a unique colour/hatching
combination.  An example if its use is in Figure 2A-C of
Scofield et al. (2012).
}
\examples{
## create table of random membership data
n.sites <- 5
n.sources <- 10
n.samples <- 160
## data frame of site-source pairs
set.seed(75333)
t <- data.frame(site = sample(n.sites, n.samples, replace = TRUE),
                source = round(runif(n.samples) * n.sources + 0.5))
## create site-by-source table
tab <- as.divtable(table(t))
membershipPlot(tab, distinguish.multiton = TRUE)
}
\references{
Scofield, D. G., Smouse, P. E., Karubian, J. and Sork, V. L. (2012)
Use of alpha, beta and gamma diversity measures to characterize seed
dispersal by animals.  \emph{American Naturalist} 180:719-732.
}

