% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nielsenTransform.R
\name{nielsenTransform}
\alias{nielsenTransform}
\title{Transform vector of squared frequencies using the method of Nielsen et al. (2003)}
\usage{
nielsenTransform(q.gg, n.g)
}
\arguments{
\item{q.gg}{Vector of squared frequencies}

\item{n.g}{Vector with group size for each element of \code{q.gg}}
}
\value{
\code{q.gg} vector with the Nielsen et al. transform applied
}
\description{
Transform vector of squared frequencies using the method of Nielsen et al. (2003)
}
\references{
Nielsen, R., Tarpy, D. R. and Reeve, H. K. (2003) Estimating effective
paternity number in social insects and the effective number of alleles in
a population.  \emph{Molecular Ecology} 12:3157-3164.
}

