% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversityTests-divtable.R
\name{print.alpha_contrast_test}
\alias{print.alpha_contrast_test}
\title{Print the result of \code{alphaContrastTest}}
\usage{
\method{print}{alpha_contrast_test}(result, ...)
}
\description{
Print the result of \code{alphaContrastTest}
}

